# Optimize L2 norm computation with optional vectorization. 